- new worldgen

+++ lib/FastNoiseLite.cs
@@ -51,8 +51,8 @@ using System;
 using System.Runtime.CompilerServices;
 
 // Switch between using floats or doubles for input position
-using FNLfloat = System.Single;
-//using FNLfloat = System.Double;
+//using FNLfloat = System.Single;
+using FNLfloat = System.Double;
 
 public class FastNoiseLite
 {
